Completed the app

+++ client/src/components/WorkoutDetails.jsx
@@ -3,14 +3,24 @@ import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
 import BootstrapModal from "./home/BootstrapModal.jsx";
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
+import { useAuthContext } from "../hooks/userAuthContext";
+
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
 
   const handleClick = async () => {
+    if (!user) {
+      return;
+    }
     const response = await fetch(
       "http://localhost:3000/api/workouts/" + workout._id,
       {
         method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: `Bearer ${user.token}`,
+        },
       }
     );
     const json = await response.json();
